+++ package.json
@@ -46,7 +46,7 @@
     "@grpc/proto-loader": "0.7.7",
     "@protobuf-ts/grpc-transport": "2.9.0",
     "@protobuf-ts/runtime-rpc": "2.9.0",
-    "@streamdal/protos": "^0.0.113",
+    "@streamdal/protos": "^0.0.118",
     "@types/rwlock": "^5.0.3",
     "@types/sinon": "^10.0.16",
     "dotenv": "^16.3.1",
